ds18b20: move wrapper to libs
